update examples and bench to account for preferred way to call moved init_cfg tools
