+++ bench/lennard_jones/espressopp/espressopp_lennard_jones.py
@@ -41,7 +41,7 @@ import logging
 from espressopp import Real3D, Int3D
 from espressopp.tools import lammps
 from espressopp.tools import decomp
-from espressopp.tools.init_cfg import lattice
+from espressopp.tools import lattice
 from espressopp.tools import timers
 
 # logging.getLogger("Storage").setLevel(logging.INFO)
